LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/presentation-markup/tables/table-default-styles-001-expected.txt


PASS Default math-style value on mtable
PASS Default display value on mtable
PASS Default display value on mtr
PASS Default display value on mtd
PASS Default text-align value on mtd
FAIL Default padding-left value on mtd assert_equals: expected "10px" but got "5.737305px"
FAIL Default padding-right value on mtd assert_equals: expected "10px" but got "5.737305px"
FAIL Default padding-top value on mtd assert_equals: expected "10px" but got "5.737305px"
FAIL Default padding-bottom value on mtd assert_equals: expected "10px" but got "5.737305px"
FAIL Bounding box is the same as mrow + mtd default padding assert_equals: expected 40 but got 31.46875
